\name{set_decimal}
\alias{set_decimal}
\title{Set a numeric (\code{x}) to (\code{k}) decimal places}
\usage{
set_decimal(x,k)
}
\arguments{
 \item{x}{Numeric object to be formated}
 \item{k}{Integer representing the number of decimal places}
}
\value{
  Formatted numeric number
}
\description{
Set a numeric (\code{x}) to (\code{k}) decimal places
}
\examples{
set_decimal(1.7511,2) # 1.75
}
