\name{split_str}
\alias{split_str}
\title{ Break comma separated string into vector-like string}
\usage{
split_str(s)
}
\arguments{
 \item{s}{comma separated string to be split}
}
\description{
 Break comma separated string into vector-like string
}
\examples{
split_str("A,B,C")  # "('A','B','C')"
}
