\name{use_ea_github}
\alias{use_ea_github}
\title{Install package from \code{eamoakohene} repository }
\usage{
use_ea_github(x)
}
\arguments{
 \item{x}{name of the repository}
}
\description{
Install package from \code{eamoakohene} repository
}
\examples{
use_ea_github('onsR2') #package for downloading data from UK ONS website
}
