\name{use_package}
\alias{use_package}
\title{Install an R package if not already installed and load otherwise just load}
\usage{
 use_package(p)
}
\arguments{
  \item{p}{The R package (library) to be loaded or installed}
}
\description{
 Install an R package if not already installed and load otherwise just load
}
\examples{
 use_package(beamaColours)
}
